+++ src/exit_codes.rs
@@ -1,6 +1,7 @@
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum ExitCode {
     Success,
+    HasResults(bool),
     GeneralError,
     KilledBySigint,
 }
@@ -9,6 +10,7 @@ impl From<ExitCode> for i32 {
     fn from(code: ExitCode) -> Self {
         match code {
             ExitCode::Success => 0,
+            ExitCode::HasResults(has_results) => !has_results as i32,
             ExitCode::GeneralError => 1,
             ExitCode::KilledBySigint => 130,
         }
@@ -17,12 +19,12 @@ impl From<ExitCode> for i32 {
 
 impl ExitCode {
     fn is_error(self) -> bool {
-        self != ExitCode::Success
+        i32::from(self) != 0
     }
 }
 
-pub fn merge_exitcodes(results: &[ExitCode]) -> ExitCode {
-    if results.iter().any(|&c| ExitCode::is_error(c)) {
+pub fn merge_exitcodes(results: impl IntoIterator<Item = ExitCode>) -> ExitCode {
+    if results.into_iter().any(ExitCode::is_error) {
         return ExitCode::GeneralError;
     }
     ExitCode::Success
@@ -34,38 +36,38 @@ mod tests {
 
     #[test]
     fn success_when_no_results() {
-        assert_eq!(merge_exitcodes(&[]), ExitCode::Success);
+        assert_eq!(merge_exitcodes([]), ExitCode::Success);
     }
 
     #[test]
     fn general_error_if_at_least_one_error() {
         assert_eq!(
-            merge_exitcodes(&[ExitCode::GeneralError]),
+            merge_exitcodes([ExitCode::GeneralError]),
             ExitCode::GeneralError
         );
         assert_eq!(
-            merge_exitcodes(&[ExitCode::KilledBySigint]),
+            merge_exitcodes([ExitCode::KilledBySigint]),
             ExitCode::GeneralError
         );
         assert_eq!(
-            merge_exitcodes(&[ExitCode::KilledBySigint, ExitCode::Success]),
+            merge_exitcodes([ExitCode::KilledBySigint, ExitCode::Success]),
             ExitCode::GeneralError
         );
         assert_eq!(
-            merge_exitcodes(&[ExitCode::Success, ExitCode::GeneralError]),
+            merge_exitcodes([ExitCode::Success, ExitCode::GeneralError]),
             ExitCode::GeneralError
         );
         assert_eq!(
-            merge_exitcodes(&[ExitCode::GeneralError, ExitCode::KilledBySigint]),
+            merge_exitcodes([ExitCode::GeneralError, ExitCode::KilledBySigint]),
             ExitCode::GeneralError
         );
     }
 
     #[test]
     fn success_if_no_error() {
-        assert_eq!(merge_exitcodes(&[ExitCode::Success]), ExitCode::Success);
+        assert_eq!(merge_exitcodes([ExitCode::Success]), ExitCode::Success);
         assert_eq!(
-            merge_exitcodes(&[ExitCode::Success, ExitCode::Success]),
+            merge_exitcodes([ExitCode::Success, ExitCode::Success]),
             ExitCode::Success
         );
     }