+++ tests/tests.rs
@@ -4,6 +4,7 @@ use std::fs;
 use std::io::Write;
 use std::path::Path;
 use std::time::{Duration, SystemTime};
+use test_case::test_case;
 
 use normpath::PathExt;
 use regex::escape;
@@ -495,6 +496,78 @@ fn test_gitignore_and_fdignore() {
         ignored-by-fdignore
         ignored-by-gitignore
         ignored-by-both",
+    );
+}
+
+/// Ignore parent ignore files (--no-ignore-parent)
+#[test]
+fn test_no_ignore_parent() {
+    let dirs = &["inner"];
+    let files = &[
+        "inner/parent-ignored",
+        "inner/child-ignored",
+        "inner/not-ignored",
+    ];
+    let te = TestEnv::new(dirs, files);
+
+    // Ignore 'parent-ignored' in root
+    fs::File::create(te.test_root().join(".gitignore"))
+        .unwrap()
+        .write_all(b"parent-ignored")
+        .unwrap();
+    // Ignore 'child-ignored' in inner
+    fs::File::create(te.test_root().join("inner/.gitignore"))
+        .unwrap()
+        .write_all(b"child-ignored")
+        .unwrap();
+
+    te.assert_output_subdirectory("inner", &[], "not-ignored");
+
+    te.assert_output_subdirectory(
+        "inner",
+        &["--no-ignore-parent"],
+        "parent-ignored
+        not-ignored",
+    );
+}
+
+/// Ignore parent ignore files (--no-ignore-parent) with an inner git repo
+#[test]
+fn test_no_ignore_parent_inner_git() {
+    let dirs = &["inner"];
+    let files = &[
+        "inner/parent-ignored",
+        "inner/child-ignored",
+        "inner/not-ignored",
+    ];
+    let te = TestEnv::new(dirs, files);
+
+    // Make the inner folder also appear as a git repo
+    fs::create_dir_all(te.test_root().join("inner/.git")).unwrap();
+
+    // Ignore 'parent-ignored' in root
+    fs::File::create(te.test_root().join(".gitignore"))
+        .unwrap()
+        .write_all(b"parent-ignored")
+        .unwrap();
+    // Ignore 'child-ignored' in inner
+    fs::File::create(te.test_root().join("inner/.gitignore"))
+        .unwrap()
+        .write_all(b"child-ignored")
+        .unwrap();
+
+    te.assert_output_subdirectory(
+        "inner",
+        &[],
+        "not-ignored
+        parent-ignored",
+    );
+
+    te.assert_output_subdirectory(
+        "inner",
+        &["--no-ignore-parent"],
+        "not-ignored
+        parent-ignored",
     );
 }
 
@@ -1346,6 +1419,48 @@ fn test_exec_batch() {
     }
 }
 
+#[test]
+fn test_exec_batch_with_limit() {
+    // TODO Test for windows
+    if cfg!(windows) {
+        return;
+    }
+
+    let te = TestEnv::new(DEFAULT_DIRS, DEFAULT_FILES);
+
+    te.assert_output(
+        &["foo", "--batch-size", "0", "--exec-batch", "echo", "{}"],
+        "a.foo one/b.foo one/two/C.Foo2 one/two/c.foo one/two/three/d.foo one/two/three/directory_foo",
+    );
+
+    let output = te.assert_success_and_get_output(
+        ".",
+        &["foo", "--batch-size=2", "--exec-batch", "echo", "{}"],
+    );
+    let stdout = String::from_utf8_lossy(&output.stdout);
+
+    for line in stdout.lines() {
+        assert_eq!(2, line.split_whitespace().count());
+    }
+
+    let mut paths: Vec<_> = stdout
+        .lines()
+        .flat_map(|line| line.split_whitespace())
+        .collect();
+    paths.sort_unstable();
+    assert_eq!(
+        &paths,
+        &[
+            "a.foo",
+            "one/b.foo",
+            "one/two/C.Foo2",
+            "one/two/c.foo",
+            "one/two/three/d.foo",
+            "one/two/three/directory_foo"
+        ],
+    );
+}
+
 /// Shell script execution (--exec) with a custom --path-separator
 #[test]
 fn test_exec_with_separator() {
@@ -1423,6 +1538,19 @@ fn test_exec_with_separator() {
         &["--path-separator=#", "e1", "--exec", "printf", "%s.%s\n"],
         "e1 e2.",
     );
+}
+
+/// Non-zero exit code (--quiet)
+#[test]
+fn test_quiet() {
+    let dirs = &[];
+    let files = &["a.foo", "b.foo"];
+    let te = TestEnv::new(dirs, files);
+
+    te.assert_output(&["-q"], "");
+    te.assert_output(&["--quiet"], "");
+    te.assert_output(&["--has-results"], "");
+    te.assert_failure_with_error(&["--quiet", "c.foo"], "")
 }
 
 /// Literal search (--fixed-strings)
@@ -1567,9 +1695,22 @@ fn create_file_with_modified<P: AsRef<Path>>(path: P, duration_in_secs: u64) {
     filetime::set_file_times(&path, ft, ft).expect("time modification failed");
 }
 
+#[cfg(test)]
+fn remove_symlink<P: AsRef<Path>>(path: P) {
+    #[cfg(unix)]
+    fs::remove_file(path).expect("remove symlink");
+
+    // On Windows, symlinks remember whether they point to files or directories, so try both
+    #[cfg(windows)]
+    fs::remove_file(path.as_ref())
+        .or_else(|_| fs::remove_dir(path.as_ref()))
+        .expect("remove symlink");
+}
+
 #[test]
 fn test_modified_relative() {
     let te = TestEnv::new(&[], &[]);
+    remove_symlink(te.test_root().join("symlink"));
     create_file_with_modified(te.test_root().join("foo_0_now"), 0);
     create_file_with_modified(te.test_root().join("bar_1_min"), 60);
     create_file_with_modified(te.test_root().join("foo_10_min"), 600);
@@ -1607,8 +1748,9 @@ fn change_file_modified<P: AsRef<Path>>(path: P, iso_date: &str) {
 }
 
 #[test]
-fn test_modified_asolute() {
+fn test_modified_absolute() {
     let te = TestEnv::new(&[], &["15mar2018", "30dec2017"]);
+    remove_symlink(te.test_root().join("symlink"));
     change_file_modified(te.test_root().join("15mar2018"), "2018-03-15T12:00:00Z");
     change_file_modified(te.test_root().join("30dec2017"), "2017-12-30T23:59:00Z");
 
@@ -1786,6 +1928,30 @@ fn test_number_parsing_errors() {
     te.assert_failure(&["--max-buffer-time=a"]);
 
     te.assert_failure(&["--max-results=a"]);
+}
+
+#[test_case("--hidden", &["--no-hidden"] ; "hidden")]
+#[test_case("--no-ignore", &["--ignore"] ; "no-ignore")]
+#[test_case("--no-ignore-vcs", &["--ignore-vcs"] ; "no-ignore-vcs")]
+#[test_case("--follow", &["--no-follow"] ; "follow")]
+#[test_case("--absolute-path", &["--relative-path"] ; "absolute-path")]
+#[test_case("-u", &["--ignore"] ; "u")]
+#[test_case("-uu", &["--ignore", "--no-hidden"] ; "uu")]
+fn test_opposing(flag: &str, opposing_flags: &[&str]) {
+    let te = TestEnv::new(DEFAULT_DIRS, DEFAULT_FILES);
+
+    let mut flags = vec![flag];
+    flags.extend_from_slice(opposing_flags);
+    let out_no_flags = te.assert_success_and_get_output(".", &[]);
+    let out_opposing_flags = te.assert_success_and_get_output(".", &flags);
+
+    assert_eq!(
+        out_no_flags,
+        out_opposing_flags,
+        "{} should override {}",
+        opposing_flags.join(" "),
+        flag
+    );
 }
 
 /// Print error if search pattern starts with a dot and --hidden is not set